progress: move code to new mfu_progress.c/h files
